refactor(l10n): rename remaining neoThemes and palettesSubtitle locale keys
