+++ lib/l10n/app_locale_en.dart
@@ -37,15 +37,15 @@ const Map<String, dynamic> appLocaleEn = {
   AppLocale.directories: 'Directories',
   AppLocale.themes: 'Themes',
   AppLocale.systemArt: 'System Art',
-  AppLocale.neoThemesSubtitle: 'Customize system card backgrounds and logos',
-  AppLocale.neoThemesNone: 'None',
-  AppLocale.neoThemesNoneSubtitle: 'Default appearance',
-  AppLocale.neoThemesLoading: 'Loading themes...',
-  AppLocale.neoThemesError: 'Could not load themes',
-  AppLocale.neoThemesApplyTitle: 'Apply theme?',
-  AppLocale.neoThemesApplyBody:
+  AppLocale.systemArtSubtitle: 'Customize system card backgrounds and logos',
+  AppLocale.systemArtNone: 'None',
+  AppLocale.systemArtNoneSubtitle: 'Default appearance',
+  AppLocale.systemArtLoading: 'Loading themes...',
+  AppLocale.systemArtError: 'Could not load themes',
+  AppLocale.systemArtApplyTitle: 'Apply theme?',
+  AppLocale.systemArtApplyBody:
       'The theme assets will be downloaded for all systems. This may take a moment.',
-  AppLocale.neoThemesDownloading: 'Downloading theme assets...',
+  AppLocale.systemArtDownloading: 'Downloading theme assets...',
   AppLocale.about: 'About',
   AppLocale.exit: 'Exit',
   AppLocale.launcher: 'Launcher',
@@ -56,7 +56,7 @@ const Map<String, dynamic> appLocaleEn = {
   AppLocale.hideRecentCard: 'Recent Games Card',
   AppLocale.hideRecentCardSubtitle:
       'Show the recently played card in the systems grid',
-  AppLocale.palettesSubtitle: 'Personalize the appearance of your NeoStation',
+  AppLocale.themesSubtitle: 'Personalize the appearance of your NeoStation',
   AppLocale.systemTheme: 'System',
 
   AppLocale.generalSettings: 'General Settings',
