+++ lib/l10n/app_locale_id.dart
@@ -37,15 +37,15 @@ const Map<String, dynamic> appLocaleId = {
   AppLocale.directories: 'Direktori',
   AppLocale.themes: 'Tema',
   AppLocale.systemArt: 'System Art',
-  AppLocale.neoThemesSubtitle: 'Sesuaikan latar belakang dan logo kartu sistem',
-  AppLocale.neoThemesNone: 'Tidak Ada',
-  AppLocale.neoThemesNoneSubtitle: 'Tampilan default',
-  AppLocale.neoThemesLoading: 'Memuat tema...',
-  AppLocale.neoThemesError: 'Tidak dapat memuat tema',
-  AppLocale.neoThemesApplyTitle: 'Terapkan tema?',
-  AppLocale.neoThemesApplyBody:
+  AppLocale.systemArtSubtitle: 'Sesuaikan latar belakang dan logo kartu sistem',
+  AppLocale.systemArtNone: 'Tidak Ada',
+  AppLocale.systemArtNoneSubtitle: 'Tampilan default',
+  AppLocale.systemArtLoading: 'Memuat tema...',
+  AppLocale.systemArtError: 'Tidak dapat memuat tema',
+  AppLocale.systemArtApplyTitle: 'Terapkan tema?',
+  AppLocale.systemArtApplyBody:
       'Aset tema akan diunduh untuk semua sistem. Ini mungkin memerlukan beberapa saat.',
-  AppLocale.neoThemesDownloading: 'Mengunduh aset tema...',
+  AppLocale.systemArtDownloading: 'Mengunduh aset tema...',
   AppLocale.about: 'Tentang',
   AppLocale.exit: 'Keluar',
   AppLocale.launcher: 'Launcher',
@@ -57,7 +57,7 @@ const Map<String, dynamic> appLocaleId = {
   AppLocale.hideRecentCard: 'Kartu Terbaru',
   AppLocale.hideRecentCardSubtitle:
       'Tampilkan kartu game terbaru di grid sistem',
-  AppLocale.palettesSubtitle: 'Personalisasi tampilan NeoStation',
+  AppLocale.themesSubtitle: 'Personalisasi tampilan NeoStation',
   AppLocale.systemTheme: 'Tema Sistem',
 
   AppLocale.generalSettings: 'Pengaturan Umum',
