+++ lib/l10n/app_locale_ja.dart
@@ -37,14 +37,14 @@ const Map<String, dynamic> appLocaleJa = {
   AppLocale.directories: 'ディレクトリ',
   AppLocale.themes: 'テーマ',
   AppLocale.systemArt: 'System Art',
-  AppLocale.neoThemesSubtitle: 'システムカードの背景とロゴをカスタマイズ',
-  AppLocale.neoThemesNone: 'なし',
-  AppLocale.neoThemesNoneSubtitle: 'デフォルトの外観',
-  AppLocale.neoThemesLoading: 'テーマを読み込み中...',
-  AppLocale.neoThemesError: 'テーマを読み込めませんでした',
-  AppLocale.neoThemesApplyTitle: 'テーマを適用しますか？',
-  AppLocale.neoThemesApplyBody: 'すべてのシステムのテーマアセットがダウンロードされます。少し時間がかかる場合があります。',
-  AppLocale.neoThemesDownloading: 'テーマアセットをダウンロード中...',
+  AppLocale.systemArtSubtitle: 'システムカードの背景とロゴをカスタマイズ',
+  AppLocale.systemArtNone: 'なし',
+  AppLocale.systemArtNoneSubtitle: 'デフォルトの外観',
+  AppLocale.systemArtLoading: 'テーマを読み込み中...',
+  AppLocale.systemArtError: 'テーマを読み込めませんでした',
+  AppLocale.systemArtApplyTitle: 'テーマを適用しますか？',
+  AppLocale.systemArtApplyBody: 'すべてのシステムのテーマアセットがダウンロードされます。少し時間がかかる場合があります。',
+  AppLocale.systemArtDownloading: 'テーマアセットをダウンロード中...',
   AppLocale.about: 'NeoStationについて',
   AppLocale.exit: '終了',
   AppLocale.launcher: 'ランチャー',
@@ -54,7 +54,7 @@ const Map<String, dynamic> appLocaleJa = {
   AppLocale.systemsSettingsSubtitle: 'ライブラリのシステムを表示または非表示にする',
   AppLocale.hideRecentCard: '最近のゲームカード',
   AppLocale.hideRecentCardSubtitle: 'システムグリッドに最近プレイしたカードを表示する',
-  AppLocale.palettesSubtitle: 'NeoStationの外観をカスタマイズ',
+  AppLocale.themesSubtitle: 'NeoStationの外観をカスタマイズ',
   AppLocale.systemTheme: 'システムテーマ',
 
   AppLocale.generalSettings: '全般設定',
