+++ lib/l10n/app_locale_ru.dart
@@ -37,15 +37,15 @@ const Map<String, dynamic> appLocaleRu = {
   AppLocale.directories: 'Директории',
   AppLocale.themes: 'Темы',
   AppLocale.systemArt: 'System Art',
-  AppLocale.neoThemesSubtitle: 'Настройте фоны и логотипы карточек систем',
-  AppLocale.neoThemesNone: 'Нет',
-  AppLocale.neoThemesNoneSubtitle: 'Внешний вид по умолчанию',
-  AppLocale.neoThemesLoading: 'Загрузка тем...',
-  AppLocale.neoThemesError: 'Не удалось загрузить темы',
-  AppLocale.neoThemesApplyTitle: 'Применить тему?',
-  AppLocale.neoThemesApplyBody:
+  AppLocale.systemArtSubtitle: 'Настройте фоны и логотипы карточек систем',
+  AppLocale.systemArtNone: 'Нет',
+  AppLocale.systemArtNoneSubtitle: 'Внешний вид по умолчанию',
+  AppLocale.systemArtLoading: 'Загрузка тем...',
+  AppLocale.systemArtError: 'Не удалось загрузить темы',
+  AppLocale.systemArtApplyTitle: 'Применить тему?',
+  AppLocale.systemArtApplyBody:
       'Ассеты темы будут загружены для всех систем. Это может занять некоторое время.',
-  AppLocale.neoThemesDownloading: 'Загрузка ассетов темы...',
+  AppLocale.systemArtDownloading: 'Загрузка ассетов темы...',
   AppLocale.about: 'О программе',
   AppLocale.exit: 'Выход',
   AppLocale.launcher: 'Лаунчер',
@@ -57,7 +57,7 @@ const Map<String, dynamic> appLocaleRu = {
   AppLocale.hideRecentCard: 'Карточка недавних игр',
   AppLocale.hideRecentCardSubtitle:
       'Показать карточку недавних игр в сетке систем',
-  AppLocale.palettesSubtitle: 'Персонализируйте внешний вид вашей NeoStation',
+  AppLocale.themesSubtitle: 'Персонализируйте внешний вид вашей NeoStation',
   AppLocale.systemTheme: 'Системная',
 
   AppLocale.generalSettings: 'Общие настройки',
