+++ lib/l10n/app_locale_zh.dart
@@ -37,14 +37,14 @@ const Map<String, dynamic> appLocaleZh = {
   AppLocale.directories: '目录',
   AppLocale.themes: '主题',
   AppLocale.systemArt: 'System Art',
-  AppLocale.neoThemesSubtitle: '自定义系统卡片背景和标志',
-  AppLocale.neoThemesNone: '无',
-  AppLocale.neoThemesNoneSubtitle: '默认外观',
-  AppLocale.neoThemesLoading: '加载主题中...',
-  AppLocale.neoThemesError: '无法加载主题',
-  AppLocale.neoThemesApplyTitle: '应用主题？',
-  AppLocale.neoThemesApplyBody: '将为所有系统下载主题资源，这可能需要一些时间。',
-  AppLocale.neoThemesDownloading: '正在下载主题资源...',
+  AppLocale.systemArtSubtitle: '自定义系统卡片背景和标志',
+  AppLocale.systemArtNone: '无',
+  AppLocale.systemArtNoneSubtitle: '默认外观',
+  AppLocale.systemArtLoading: '加载主题中...',
+  AppLocale.systemArtError: '无法加载主题',
+  AppLocale.systemArtApplyTitle: '应用主题？',
+  AppLocale.systemArtApplyBody: '将为所有系统下载主题资源，这可能需要一些时间。',
+  AppLocale.systemArtDownloading: '正在下载主题资源...',
   AppLocale.about: '关于',
   AppLocale.exit: '退出',
   AppLocale.launcher: '启动器',
@@ -54,7 +54,7 @@ const Map<String, dynamic> appLocaleZh = {
   AppLocale.systemsSettingsSubtitle: '显示或隐藏库中的系统',
   AppLocale.hideRecentCard: '最近游戏卡片',
   AppLocale.hideRecentCardSubtitle: '在系统网格中显示最近游玩卡片',
-  AppLocale.palettesSubtitle: '个性化您的 NeoStation 外观',
+  AppLocale.themesSubtitle: '个性化您的 NeoStation 外观',
   AppLocale.systemTheme: '系统默认',
 
   AppLocale.generalSettings: '常规设置',
