+++ lib/l10n/app_locale_zh_hant.dart
@@ -37,14 +37,14 @@ const Map<String, dynamic> appLocaleZhHant = {
   AppLocale.directories: '目錄',
   AppLocale.themes: '主題',
   AppLocale.systemArt: 'System Art',
-  AppLocale.neoThemesSubtitle: '自訂系統卡片背景和標誌',
-  AppLocale.neoThemesNone: '無',
-  AppLocale.neoThemesNoneSubtitle: '預設外觀',
-  AppLocale.neoThemesLoading: '正在載入主題...',
-  AppLocale.neoThemesError: '無法載入主題',
-  AppLocale.neoThemesApplyTitle: '套用主題？',
-  AppLocale.neoThemesApplyBody: '將為所有系統下載主題資源，這可能需要一些時間。',
-  AppLocale.neoThemesDownloading: '正在下載主題資源...',
+  AppLocale.systemArtSubtitle: '自訂系統卡片背景和標誌',
+  AppLocale.systemArtNone: '無',
+  AppLocale.systemArtNoneSubtitle: '預設外觀',
+  AppLocale.systemArtLoading: '正在載入主題...',
+  AppLocale.systemArtError: '無法載入主題',
+  AppLocale.systemArtApplyTitle: '套用主題？',
+  AppLocale.systemArtApplyBody: '將為所有系統下載主題資源，這可能需要一些時間。',
+  AppLocale.systemArtDownloading: '正在下載主題資源...',
   AppLocale.about: '關於',
   AppLocale.exit: '離開',
   AppLocale.launcher: '啟動器',
@@ -54,7 +54,7 @@ const Map<String, dynamic> appLocaleZhHant = {
   AppLocale.systemsSettingsSubtitle: '顯示或隱藏收藏庫中的系統',
   AppLocale.hideRecentCard: '最近遊戲卡片',
   AppLocale.hideRecentCardSubtitle: '在系統網格中顯示最近遊玩卡片',
-  AppLocale.palettesSubtitle: '個人化您的 NeoStation 外觀',
+  AppLocale.themesSubtitle: '個人化您的 NeoStation 外觀',
   AppLocale.systemTheme: '系統預設',
 
   AppLocale.generalSettings: '一般設定',
